backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hutanmihai <hutanmihai29@gmail.com>"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "3.12.2"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = { extras = ["postgresql-asyncpg"], version = "^2.0.29" }
pydantic-settings = "^2.2.1"
pyjwt = "^2.8.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pydantic = {extras = ["email"], version = "^2.6.4"}


[tool.poetry.group.dev.dependencies]
black = "24.2.0"
poethepoet = "^0.25.0"
alembic = "^1.13.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
[tool.poe.tasks.create_migration]
# This task will create migration
help = "Create migration"
cmd = "alembic revision --autogenerate"
envfile = ".env"

[tool.poe.tasks.upgrade_schema]
# This task will upgrade the database schema
help = "Upgrade database schema"
cmd = "alembic upgrade head"
envfile = ".env"

[tool.poe.tasks.downgrade_schema]
# This task will upgrade the database schema
help = "Downgrade database schema"
cmd = "alembic downgrade head"
envfile = ".env"
